Implement some interactions using the clickable image resource
Using the demo rocket images, add some interactions when they are clicked.

Add a mouse area for the entire background, so that when clicked it will
reset the ui to the original layout

When rocket 1 is clicked, move it 5 pixels right.
When rocket 2 is clicked, rotate it by 5 degrees
When rocket 3 is clicked, rotate it by 5 degrees and reduce its scale by 5%